+++ src/packets/packet_generator.ts
@@ -14,7 +14,7 @@ import {
   PageColorType,
   BitmapColorMode,
 } from ".";
-import { EncodedImage, ImageEncoder } from "../image_encoder";
+import { EncodedImage, ImageEncoder } from "../utils";
 import { Utils, Validators } from "../utils";
 
 export interface ImagePacketsGenerateOptions {
@@ -461,5 +461,9 @@ export class PacketGenerator {
 
   public static getPaperInfo(): NiimbotPacket {
     return this.mapped(TX.GetPaperInfo);
+  }
+
+  public static getPrinterCapabilities(): NiimbotPacket {
+    return this.mapped(TX.GetPrinterCapabilities);
   }
 }