+++ src/printer_models.ts
@@ -1,7 +1,7 @@
 /* AUTO-GENERATED FILE. DO NOT EDIT! */
 /* use 'npm run gen-printer-models' to generate */
 
-import { PrintDirection } from "./image_encoder";
+import { PrintDirection } from "./utils/image_encoder";
 import { LabelType as LT } from "./packets";
 
 /** @category Printer model library */